- Cleaning up and refactoring.

+++ DLL Solution/Bot/SpelunkBot/WalkAction.h
@@ -4,13 +4,11 @@
 class WalkAction : public IMovementAction
 {
 private:
-	double _distance;
+	int _distance;
 	bool _usingPixelCoords;
 	bool _running;
 
 public:
-	//WalkAction(IBot *bot, bool goingRight, double distance, bool usingPixelCoords);
-	//WalkAction(IBot *bot, bool goingRight, bool pressRun, double distance, bool usingPixelCoords);
 	WalkAction(IBot *bot, int dist);
 	WalkAction(IBot *bot, int dist, bool run);
 	void AddDistance(int distance);
